feature: useUsers moved to home
