+++ src/views/Home/hooks/useUsers.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getUsersList } from "..";
+import { getUsersList } from "../../../redux/slices/users";
 
 export const useUsers = () => {
   const { usersData } = useSelector((state) => state.users);
